third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-invoke-in-sync-event-expected.txt

This is a testharness.js-based test.
[FAIL] await a stable state and sync event handlers
  assert_equals: networkState after src removed expected 0 but got 3
Harness: the test ran to completion.
